refactor: data types for video items

+++ app/src/main/java/dev/saurabhmishra/exoplayersample/database/entity/VideoEntity.kt
@@ -9,7 +9,7 @@ data class VideoEntity(
     val videoUrl: String,
     val title: String,
     val views: Long,
-    val likeCount: Int,
-    val dislikeCount: Int,
+    val likeCount: Long,
+    val dislikeCount: Long,
     val uploadedTimeStamp: Long
 )
